% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_score.R
\name{entropy_score}
\alias{entropy_score}
\title{Entropy Uncertainty Score}
\usage{
entropy_score(score_matrix)
}
\arguments{
\item{score_matrix}{SingleR score matrix}
}
\description{
Calculates uncertainty from
SingleR score matrix.
}
