% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_labels.R
\name{match_labels}
\alias{match_labels}
\title{Match Cell Labels to Marker Ontology}
\usage{
match_labels(label, marker_names)
}
\description{
Automatically matches predicted
labels to available marker sets.
}
