% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_cell.R
\name{explain_cell}
\alias{explain_cell}
\title{Explain Cell Annotation Confidence}
\usage{
explain_cell(object, cell_id)
}
\arguments{
\item{object}{Seurat object}

\item{cell_id}{cell barcode}
}
\description{
Provides interpretable explanation
for confidence scores.
}
