% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_cell.R
\name{explain_cell}
\alias{explain_cell}
\title{Explain Cell Annotation Confidence}
\usage{
explain_cell(object, cell_id)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object.}

\item{cell_id}{Cell barcode or column name.}
}
\value{
A character vector containing the interpretation.
}
\description{
Provides an interpretable explanation for the confidence score
assigned to an individual cell.
}
\examples{
\dontrun{
## Assume 'object' has already been processed with cell_certify().
explain_cell(
  object,
  cell_id = "Cell1"
)
}

}
