% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_confidence.R
\name{explain_confidence}
\alias{explain_confidence}
\title{Explain Confidence Attribution}
\usage{
explain_confidence(object, cell_id)
}
\arguments{
\item{object}{Seurat object}

\item{cell_id}{Cell barcode}
}
\value{
Character vector
}
\description{
Explains why a cell received
its confidence score.
}
