% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_score.R
\name{marker_score}
\alias{marker_score}
\title{Marker Consistency Score}
\usage{
marker_score(object, markers, label_column = "predicted_label")
}
\arguments{
\item{object}{Seurat object}

\item{markers}{named marker list}

\item{label_column}{metadata column}
}
\description{
Calculates marker-based confidence score.
}
