% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_score.R
\name{marker_score}
\alias{marker_score}
\title{Marker Consistency Score}
\usage{
marker_score(object, markers, label_column = "predicted_label")
}
\arguments{
\item{object}{Seurat object}

\item{markers}{Named marker gene list}

\item{label_column}{Metadata label column}
}
\value{
Numeric vector
}
\description{
Uses UCell enrichment scoring
for robust marker evaluation.
}
