% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_score.R
\name{neighbor_score}
\alias{neighbor_score}
\title{Neighbor Agreement Score}
\usage{
neighbor_score(
  object,
  reduction = "pca",
  dims = NULL,
  k = 10,
  label_column = "predicted_label"
)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Dimensional reduction method}

\item{dims}{PCA dimensions}

\item{k}{Number of neighbors}

\item{label_column}{Metadata label column}
}
\value{
Numeric vector
}
\description{
Computes local label agreement.
}
